moving stuff to common
